[sprint_2] создание Proxy вынесено в утилиты
