+++ src/app/modules/event-bus.ts
@@ -1,4 +1,11 @@
-export default class EventBus {
+export enum EVENTS {
+  INIT = 'init',
+  FLOW_CDM = 'flow:component-did-mount',
+  FLOW_CDU = 'flow:component-did-update',
+  FLOW_RENDER = 'flow:render'
+}
+
+export class EventBus {
   listeners: { [key: string]: Array<() => void> };
 
   constructor() {
